Sprint 19 Playlistmaker moved to Fragments and used BottomNavigateView
