+++ app/src/main/AndroidManifest.xml
@@ -8,6 +8,7 @@
     <application
         android:name=".utils.App"
         android:allowBackup="true"
+
         android:dataExtractionRules="@xml/data_extraction_rules"
         android:fullBackupContent="@xml/backup_rules"
         android:icon="@mipmap/ic_launcher"
@@ -18,41 +19,19 @@
         android:usesCleartextTraffic="true"
         tools:targetApi="31">
         <activity
-            android:name=".root.presentation.RootActivity"
-            android:exported="true">
+            android:name=".root.presentation.ui.RootActivity"
+            android:exported="true"
+            android:windowSoftInputMode = "adjustNothing">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
 
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
+
         <activity
             android:name=".player.presentation.ui.PlayerActivity"
             android:exported="false" />
-        <activity
-            android:name=".search.presentation.ui.SearchActivity"
-            android:exported="false">
-            <meta-data
-                android:name="android.app.lib_name"
-                android:value="" />
-        </activity>
-        <activity
-            android:name=".mediateka.presentation.ui.MediaActivity"
-            android:exported="false">
-            <meta-data
-                android:name="android.app.lib_name"
-                android:value="" />
-        </activity>
-        <activity
-            android:name=".main.presentation.ui.MainActivity"
-            android:exported="true">
-            <meta-data
-                android:name="android.app.lib_name"
-                android:value="" />
-        </activity>
-        <activity
-            android:name=".settings.presentation.ui.SettingsActivity"
-            android:exported="false" />
-    </application>
 
+    </application>
 </manifest>
