--- app/src/main/java/LocalStorage.kt
@@ -1,14 +1,0 @@
-/*
-package com.practicum.playlistmaker.search.data.storage
-
-import android.content.SharedPreferences
-
-class LocalStorage(private val sharedPreferences: SharedPreferences) {
-    fun save(){
-
-    }
-
-    fun get(){
-
-    }
-}*/
