--- app/src/main/java/com/practicum/playlistmaker/di/MainModule.kt
@@ -1,7 +1,0 @@
-package com.practicum.playlistmaker.di
-
-import org.koin.dsl.module
-
-val mainModule = module{
-
-}
