--- app/src/main/java/com/practicum/playlistmaker/main/presentation/MainViewModel.kt
@@ -1,7 +1,0 @@
-package com.practicum.playlistmaker.main.presentation
-
-import androidx.lifecycle.ViewModel
-
-class MainViewModel:ViewModel() {
-
-}
