--- app/src/main/java/com/practicum/playlistmaker/mediateka/presentation/ui/MediaActivity.kt
@@ -1,32 +1,0 @@
-package com.practicum.playlistmaker.mediateka.presentation.ui
-
-import android.os.Bundle
-import androidx.appcompat.app.AppCompatActivity
-import com.google.android.material.tabs.TabLayoutMediator
-import com.practicum.playlistmaker.R
-import com.practicum.playlistmaker.databinding.ActivityMediaBinding
-
-class MediaActivity : AppCompatActivity() {
-    private lateinit var binding: ActivityMediaBinding
-    private lateinit var tabMediator: TabLayoutMediator
-
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        binding = ActivityMediaBinding.inflate(layoutInflater)
-        setContentView(binding.root)
-
-
-        binding.toolbarActivityMediateka.setNavigationOnClickListener {
-            this.finish()
-        }
-
-        binding.viewPager.adapter = MediatekaViewPagerAdapter(supportFragmentManager, lifecycle)
-        tabMediator = TabLayoutMediator(binding.tabLayout, binding.viewPager) { tab, position ->
-            when (position) {
-                0 -> tab.text = getString(R.string.favourites_tab_text)
-                1 -> tab.text = getString(R.string.playlists_tab_text)
-            }
-        }
-        tabMediator.attach()
-    }
-}
