+++ app/src/main/java/com/practicum/playlistmaker/mediateka/presentation/ui/MediatekaFragment.kt
@@ -1,0 +1,38 @@
+package com.practicum.playlistmaker.mediateka.presentation.ui
+
+import android.os.Bundle
+import android.view.LayoutInflater
+import android.view.View
+import android.view.ViewGroup
+import androidx.fragment.app.Fragment
+import com.google.android.material.tabs.TabLayoutMediator
+import com.practicum.playlistmaker.R
+import com.practicum.playlistmaker.databinding.FragmentMediatekaBinding
+
+
+class MediatekaFragment: Fragment() {
+    private lateinit var binding: FragmentMediatekaBinding
+    private lateinit var tabMediator: TabLayoutMediator
+
+    override fun onCreateView(
+        inflater: LayoutInflater,
+        container: ViewGroup?,
+        savedInstanceState: Bundle?
+    ): View? {
+        binding = FragmentMediatekaBinding.inflate(inflater, container, false)
+        return binding.root
+    }
+
+    override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
+        super.onViewCreated(view, savedInstanceState)
+
+        binding.viewPager.adapter = MediatekaViewPagerAdapter(childFragmentManager, lifecycle)
+        tabMediator = TabLayoutMediator(binding.tabLayout, binding.viewPager) { tab, position ->
+            when (position) {
+                0 -> tab.text = getString(R.string.favourites_tab_text)
+                1 -> tab.text = getString(R.string.playlists_tab_text)
+            }
+        }
+        tabMediator.attach()
+    }
+}
