+++ app/src/main/res/drawable/bottom_navigation_view_selector.xml
@@ -1,0 +1,5 @@
+<?xml version="1.0" encoding="utf-8"?>
+<selector xmlns:android="http://schemas.android.com/apk/res/android">
+    <item android:state_checked="true" android:color="@color/yp_blue" />
+    <item android:color="?attr/colorOnPrimary"  />
+</selector>
