+++ app/src/main/res/drawable/icon_mediateka_fragment.xml
