+++ app/src/main/res/layout/fragment_search.xml
@@ -13,7 +13,6 @@
         android:layout_height="@dimen/dp50"
         android:layout_marginTop="@dimen/dp6"
         android:layout_marginBottom="@dimen/dp14"
-        app:navigationIcon="@drawable/icon_arrow_back"
         app:navigationIconTint="?android:attr/textColor">
 
         <TextView
