+++ app/src/main/res/menu/bottom_navigate_view_menu.xml
@@ -1,0 +1,17 @@
+<?xml version="1.0" encoding="utf-8"?>
+<menu xmlns:android="http://schemas.android.com/apk/res/android">
+    <item
+        android:id="@+id/searchFragment"
+        android:icon="@drawable/icon_search_fragment"
+        android:title="@string/menu_text_search"/>
+
+    <item
+        android:id="@+id/mediatekaFragment"
+        android:icon="@drawable/icon_mediateka_fragment"
+        android:title="@string/menu_text_mediateka"/>
+
+    <item
+        android:id="@+id/settingsFragment"
+        android:icon="@drawable/icon_settings"
+        android:title="@string/menu_text_settings" />
+</menu>
