+++ app/src/main/res/navigation/nav_graph.xml
@@ -3,7 +3,7 @@
     xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
     android:id="@+id/nav_graph"
-    app:startDestination="@id/searchFragment">
+    app:startDestination="@id/mediatekaFragment">
 
     <fragment
         android:id="@+id/searchFragment"
@@ -15,4 +15,9 @@
         android:name="com.practicum.playlistmaker.settings.presentation.ui.SettingsFragment"
         android:label="fragment_settings"
         tools:layout="@layout/fragment_settings" />
+    <fragment
+        android:id="@+id/mediatekaFragment"
+        android:name="com.practicum.playlistmaker.mediateka.presentation.ui.MediatekaFragment"
+        android:label="fragment_mediateka"
+        tools:layout="@layout/fragment_mediateka" />
 </navigation>
